timeutils: extract zoneinfo from timeutils.{c,h}
