+++ lib/timeutils/timeutils.h
@@ -50,13 +50,6 @@ glong timespec_diff_msec(const struct timespec *t1, const struct timespec *t2);
 glong timespec_diff_nsec(struct timespec *t1, struct timespec *t2);
 gint determine_year_for_month(gint month, const struct tm *now);
 
-typedef struct _ZoneInfo ZoneInfo;
-typedef struct _TimeZoneInfo TimeZoneInfo;
-
-gint32 time_zone_info_get_offset(const TimeZoneInfo *self, time_t stamp);
-TimeZoneInfo *time_zone_info_new(const gchar *tz);
-void time_zone_info_free(TimeZoneInfo *self);
-
 extern const char *month_names_abbrev[];
 extern const char *month_names[];
 extern const char *weekday_names_abbrev[];
