+++ lib/timeutils/zoneinfo.h
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2002-2019 Balabit
+ * Copyright (c) 1998-2019 Balázs Scheidler
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+ *
+ * As an additional exemption you are allowed to compile & link against the
+ * OpenSSL libraries as published by the OpenSSL project. See the file
+ * COPYING for details.
+ *
+ */
+
+#ifndef TIMEUTILS_ZONEINFO_H_INCLUDED
+#define TIMEUTILS_ZONEINFO_H_INCLUDED
+
+#include "syslog-ng.h"
+
+typedef struct _ZoneInfo ZoneInfo;
+typedef struct _TimeZoneInfo TimeZoneInfo;
+
+gint32 time_zone_info_get_offset(const TimeZoneInfo *self, time_t stamp);
+TimeZoneInfo *time_zone_info_new(const gchar *tz);
+void time_zone_info_free(TimeZoneInfo *self);
+
+
+#endif
