+++ modules/dbparser/groupingby.c
@@ -27,6 +27,7 @@
 #include "messages.h"
 #include "str-utils.h"
 #include "filter/filter-expr.h"
+#include "timeutils/timeutils.h"
 #include <iv.h>
 
 typedef struct _GroupingBy
