+++ modules/dbparser/pdbtool/pdbtool.c
@@ -44,6 +44,7 @@
 #include "crypto.h"
 #include "compat/openssl_support.h"
 #include "scratch-buffers.h"
+#include "timeutils/timeutils.h"
 
 #include <stdio.h>
 #include <string.h>
