+++ modules/diskq/tests/test_diskq.c
@@ -37,6 +37,7 @@
 #include "queue_utils_lib.h"
 #include "test_diskq_tools.h"
 #include "testutils.h"
+#include "timeutils/timeutils.h"
 
 #include <stdlib.h>
 #include <string.h>
