+++ tests/unit/test_zone.c
@@ -23,7 +23,7 @@
 
 #include "apphook.h"
 #include "timeutils/timeutils.h"
-#include "timeutils/timeutils.h"
+#include "timeutils/zoneinfo.h"
 #include "logstamp.h"
 #include <criterion/criterion.h>
 #include <stdlib.h>
